Move buildtime constants into an internal package.
