+++ cmd/scrum/internal/buildtime/consts.go
